# backend/go-orchestrator/docs/swagger.yaml
basePath: /api
definitions:
  github_com_skr1ms_Dev2GIS_pkg_jwt.TokenPair:
    properties:
      access_token:
        type: string
      expires_at:
        type: integer
      refresh_token:
        type: string
    type: object
  internal_auth.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_auth.LoginResponse:
    properties:
      tokens:
        $ref: '#/definitions/github_com_skr1ms_Dev2GIS_pkg_jwt.TokenPair'
      user:
        $ref: '#/definitions/internal_auth.UserResponse'
    type: object
  internal_auth.LoginUserRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_auth.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  internal_heightmap.HeightmapJob:
    properties:
      created_at:
        type: string
      error_message:
        type: string
      height:
        type: integer
      id:
        type: string
      image_url:
        type: string
      processing_time:
        type: number
      result_url:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      width:
        type: integer
    type: object
  internal_heightmap.UploadResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Dev2GIS Support
  description: API for UAV data processing and 3D height map generation
  title: UAV Heights Map API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: User login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: User Login
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: User Logout
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Token refresh request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Refresh JWT Token
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: User registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create User
      tags:
      - auth
  /api/heightmaps:
    get:
      description: Get list of all available height maps for current user
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Height Maps
      tags:
      - heightmaps
  /api/heightmaps/{id}:
    get:
      description: Retrieve height map metadata by ID
      parameters:
      - description: Height Map ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_heightmap.HeightmapJob'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Height Map by ID
      tags:
      - heightmaps
  /api/heightmaps/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a photo and generate 3D height map
      parameters:
      - description: Photo file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_heightmap.UploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload Photo for Height Map
      tags:
      - heightmaps
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"