# backend/go-orchestrator/internal/storage/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package sqlc

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type BatchHeightmapJob struct {
	ID             uuid.UUID `json:"id"`
	UserID         uuid.UUID `json:"user_id"`
	Status         string    `json:"status"`
	ResultUrl      *string   `json:"result_url"`
	OrthophotoUrl  *string   `json:"orthophoto_url"`
	Width          *int32    `json:"width"`
	Height         *int32    `json:"height"`
	ImageCount     int32     `json:"image_count"`
	ProcessedCount int32     `json:"processed_count"`
	ErrorMessage   *string   `json:"error_message"`
	ProcessingTime *float64  `json:"processing_time"`
	MergeMethod    string    `json:"merge_method"`
	GenerationMode string    `json:"generation_mode"`
	CreatedAt      time.Time `json:"created_at"`
	UpdatedAt      time.Time `json:"updated_at"`
}

type BatchImage struct {
	ID             uuid.UUID   `json:"id"`
	BatchJobID     pgtype.UUID `json:"batch_job_id"`
	ImageUrl       string      `json:"image_url"`
	HeightmapJobID pgtype.UUID `json:"heightmap_job_id"`
	Status         string      `json:"status"`
	CreatedAt      time.Time   `json:"created_at"`
}

type HeightmapJob struct {
	ID             uuid.UUID `json:"id"`
	UserID         uuid.UUID `json:"user_id"`
	ImageUrl       string    `json:"image_url"`
	ResultUrl      *string   `json:"result_url"`
	Status         string    `json:"status"`
	Width          *int32    `json:"width"`
	Height         *int32    `json:"height"`
	ErrorMessage   *string   `json:"error_message"`
	ProcessingTime *float64  `json:"processing_time"`
	CreatedAt      time.Time `json:"created_at"`
	UpdatedAt      time.Time `json:"updated_at"`
}

type User struct {
	ID           uuid.UUID `json:"id"`
	Email        string    `json:"email"`
	PasswordHash string    `json:"password_hash"`
	Name         string    `json:"name"`
	CreatedAt    time.Time `json:"created_at"`
	UpdatedAt    time.Time `json:"updated_at"`
}